fix: sim-cloud-front-distro-router.ts FTA <50
