+++ src/service/cloudfront/router/sim-cloudfront-distro-router.iso.test.ts
@@ -14,6 +14,7 @@ import { CreateDistributionCommand } from "@aws-sdk/client-cloudfront";
 import { CreateBucketCommand } from "@aws-sdk/client-s3";
 import { SimCloudFrontRegistry } from "../sim-cloud-front-registry.js";
 import type { SimAwsAccountId } from "../../aws/sim-aws-account.js";
+import { extractCloudFrontHostDistroId } from "./extract-cf-host-distro-id.js";
 
 describe("Sim CloudFront Distribution Router", () => {
   describe("route requests to distributions", () => {
@@ -250,36 +251,35 @@ describe("Sim CloudFront Distribution Router", () => {
 
   describe("extract Distribution ID from hostname", () => {
     it("matches cloudfront distro hostname alone", () => {
-      const distributionId = SimCloudFrontDistroRouter.extractHostDistroId(
+      const distributionId = extractCloudFrontHostDistroId(
         "distro123.cloudfront.net",
       );
       assertIdentical(distributionId, "DISTRO123");
     });
 
     it("matches cloudfront distro hostname prefix", () => {
-      const distributionId = SimCloudFrontDistroRouter.extractHostDistroId(
+      const distributionId = extractCloudFrontHostDistroId(
         "distro123.cloudfront.net.sim-aws.localhost",
       );
       assertIdentical(distributionId, "DISTRO123");
     });
 
     it("does not match non-CloudFront hostname", () => {
-      const distributionId = SimCloudFrontDistroRouter.extractHostDistroId(
+      const distributionId = extractCloudFrontHostDistroId(
         "distro123.foo.net.sim-aws.localhost",
       );
       assertUndefined(distributionId);
     });
 
     it("does not match non-net TLD", () => {
-      const distributionId = SimCloudFrontDistroRouter.extractHostDistroId(
+      const distributionId = extractCloudFrontHostDistroId(
         "distro123.cloudfront.com.sim-aws.localhost",
       );
       assertUndefined(distributionId);
     });
 
     it("does not match short hostname", () => {
-      const distributionId =
-        SimCloudFrontDistroRouter.extractHostDistroId("sim-aws.localhost");
+      const distributionId = extractCloudFrontHostDistroId("sim-aws.localhost");
       assertUndefined(distributionId);
     });
   });
